create stories from operation
